requirements.txt
numpy
gym
tensorflow-gpu
scipy